+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - fix generated version.py with newline and double-quotes to silence lint warning
+- add PDM pre_publish hook to block on unreleased entries in changelog or untracked/uncommitted files in git ws
 
 ## [0.1.2] - 2025-10-15
 ### Fixed